--- alembic/versions/be9552e4fb7d_cascading_playlist_track.py
@@ -1,45 +1,0 @@
-"""cascading playlist_track
-
-Revision ID: be9552e4fb7d
-Revises: 
-Create Date: 2023-05-22 23:10:59.460153
-
-"""
-from alembic import op
-import sqlalchemy as sa
-
-
-# revision identifiers, used by Alembic.
-revision = "be9552e4fb7d"
-down_revision = None
-branch_labels = None
-depends_on = None
-
-
-def upgrade() -> None:
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.drop_constraint(
-        "playlist_track_playlist_id_fkey", "playlist_track", type_="foreignkey"
-    )
-    op.create_foreign_key(
-        None,
-        "playlist_track",
-        "playlists",
-        ["playlist_id"],
-        ["playlist_id"],
-        ondelete="CASCADE",
-    )
-    # ### end Alembic commands ###
-
-
-def downgrade() -> None:
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.drop_constraint(None, "playlist_track", type_="foreignkey")
-    op.create_foreign_key(
-        "playlist_track_playlist_id_fkey",
-        "playlist_track",
-        "playlists",
-        ["playlist_id"],
-        ["playlist_id"],
-    )
-    # ### end Alembic commands ###
